requirements.txt
wxPython
pycaw
requests
packaging
python-vlc
nvda-controller-client-python
